+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/CoopLockRepairIntegrationTest.java
@@ -101,7 +101,8 @@ public class CoopLockRepairIntegrationTest {
             gcsOptions.getAppName(),
             gcsOptions.getMaxHttpRequestRetries(),
             gcsOptions.getHttpRequestConnectTimeout(),
-            gcsOptions.getHttpRequestReadTimeout());
+            gcsOptions.getHttpRequestReadTimeout(),
+            gcsOptions.getHttpRequestHeaders());
 
     GoogleCloudStorageFileSystem gcsfs =
         new GoogleCloudStorageFileSystem(