+++ util/src/test/java/com/google/cloud/hadoop/util/RetryHttpInitializerTest.java
@@ -34,6 +34,7 @@ import com.google.api.client.http.LowLevelHttpRequest;
 import com.google.api.client.http.LowLevelHttpResponse;
 import com.google.api.client.testing.http.MockHttpTransport;
 import com.google.api.client.util.Sleeper;
+import com.google.common.collect.ImmutableMap;
 import java.io.IOException;
 import org.junit.After;
 import org.junit.Before;
@@ -86,7 +87,14 @@ public class RetryHttpInitializerTest {
         return mockLowLevelRequest;
       }
     };
-    initializer = new RetryHttpInitializer(mockCredential, "foo-user-agent");
+    initializer =
+        new RetryHttpInitializer(
+            mockCredential,
+            "foo-user-agent",
+            HttpRequest.DEFAULT_NUMBER_OF_RETRIES,
+            20_000,
+            20_000,
+            ImmutableMap.of("header-key", "header=value"));
     initializer.setSleeperOverride(mockSleeper);
     requestFactory = fakeTransport.createRequestFactory(initializer);
   }
@@ -108,6 +116,7 @@ public class RetryHttpInitializerTest {
     final String authHeaderValue = "Bearer a1b2c3d4";
     final HttpRequest req = requestFactory.buildGetRequest(new GenericUrl("http://fake-url.com"));
     assertThat(req.getHeaders().getUserAgent()).isEqualTo("foo-user-agent");
+    assertThat(req.getHeaders().get("header-key")).isEqualTo("header=value");
     assertThat(req.getInterceptor()).isEqualTo(mockCredential);
 
     // Simulate the actual behavior of inserting a header for the credential.